refactor name split method into helper class
