+++ app/models/application_merits_task/opponent.rb
@@ -1,5 +1,7 @@
 module ApplicationMeritsTask
   class Opponent < ApplicationRecord
+    include NameSplitHelper
+
     belongs_to :legal_aid_application
   end
 end
